fix(github): bump version of pnpm/action-setup to v3 (#3359)
Related to #3337 

This is to upgrade GitHub action that sets up PNPM. The action has [released a new version](https://github.com/pnpm/action-setup/releases) that is developed on node `20` and this will stop warnings in the GitHub workflows.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.