fix(github): bump peter-evans/create-pull-request to v6 (#3361)
Relates to #3337 

This is to upgrade GitHub action that creates Pull Request. The action has [released a new version](https://github.com/peter-evans/create-pull-request/releases) that is developed on node 20 and this will stop warnings in the GitHub workflows.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.