feat(node): release from CI with npm provenance statements (#3330)
This PR adds `npmProvenance` to allow increasing supply-chain security for a package.

npm and pnpm support it, but Yarn still needs to, as per [npm guide](https://docs.npmjs.com/generating-provenance-statements). It's supported only for public packages.

It will be false by default.

## References

https://docs.npmjs.com/generating-provenance-statements
https://docs.npmjs.com/cli/v10/using-npm/config#provenance

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.