chore: bootstrap with NodeJS script for projen cross-platform development (#3377)
## Motivation

I was taking inspiration from https://github.com/projen/projen/pull/2761. I didn't closely follow what was available there because it's outdated, but I used a similar approach.

There are different issues related to developing `projen` on Windows. The most related ones to bootstrapping are #1529 and #2427.

To break the Windows development problem into pieces, I started with bootstraping because it is command I run the most run when developing projen. The other projen tasks can be reviewed and modified if necessary in future PRs.

I should create an issue to expose this idea, but before doing that, I'm trying to see if breaking it like that is viable and starting with bootstrapping.

## Implementation

I decided to do something similar to #2761 and use a NodeJS script. Then, I translated the existing Bash script, as of today, to NodeJS. Also, `.projenrc.js` will generate that script, similar to how it works with the existing Bash script. Windows has no CI implementation because the `build` and `package` tasks could fail. 

## Changes

## Change 1 projen.js

1. The bootstrapping script is a translation from Bash to NodeJS.
2. The critical bits are the `bootstrap` and `execCommand` functions.
3. In `execCommand`, the important part is `{ stdio: "inherit" }` to make sure to pass input and get output from the executed command by passing down the parent process streams to the child process.
4. I didn't add tests.

## Change 2 .projenrc.js

1. Replacing `projen.bash` with `projen.js`
2. Adding functions in the entrypoint `.projenrc.js`: `execCommand`  and `bootstrap`. These are only to simplify how `projen.js` is generated. Could it be better to move those two functions to another file so as not to pollute .projenrc.js?
4. I didn't add tests.

## Change 3 all files with marker

1. The marker was autoupdated to use `node ./projen.js`

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.