chore: fix missing type in building your own project docs (#3407)
Fixes
prMention missing in the MyMicroserviceProjectOptions interface
defaultReleaseBranch shouldn't be there in the example because this example should solely help with the goal of adding new prMention property, not dragging in what the default options the parent class has.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.