chore: change projen entrypoint to typescript (#3399)
Related to https://github.com/projen/projen/issues/3384

## Motivation

I wanted to add the GitHub workflow for Windows in `projen`. That would be in a `projenrc` folder to not clutter the `.projenrc.js`

I thought about using the `projenrc` root folder based on this PR's comment: https://github.com/projen/projen/pull/3377#discussion_r1496454238 . The `Projenrc` component already has optionsfor that folder that could be used for what's needed for this CI workflow. This PR is the first step to use that `Projenrc` component and change the entrypoint `.projenrc.ts` to TypeScript.

I analyzed converting the bootstrap script `projen.js` to TypeScript too but didn't find a way to solve the hashbang line. So, for now, that file is kept as JavaScript.

## Changelog in .projenrc.ts

1. Change entrypoint from `.projenrc.js` to `.projenrc.ts`.
1. Add `projenrc: true`.
1. bootstrap functions moved back to `.projenrc.ts` as `TextFile.lines` because `function.toString()` transpiles with TypeScript in an undesirable way. 
1. Delete the field `compileBeforeTest` that doesn't exist anymore.
1. Check if tasks are undefined.
1. Replace `authorUrl` with `authorAddress`. `authorUrl` is mentioned as deprecated in https://github.com/projen/projen/blob/cc06084d2990d91e7dd9f9b14b8caf4b9d38684e/src/cdk/jsii-project.ts#L590-L594
1. Remove `templates` inclusion from gitignore because that directory doesn't exist.

Tests:

1. No tests were added.

## Changelog for marker in other files

1. Marker auto-updated in all needed files.

## Migrations needed.

1. `yarn install` needs to be run before any other task because of the added requirement of `ts-node`.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.