fix(typescript): TSConfig `compilerOptions` can be optional (#3436)
Fixes #3411

Note: Still unable to make an empty `compilerOptions` (without providing one with everything explicitly set to `undefined`) due to `TypeScriptProject` forcing an override. I'll make another issue and PR for that.

Apologies to @coleHafner if they've already spent time on this

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.