fix(cli): regression cannot use named local package to create new projects (#3455)
A named local package is a package that has a local file pointer in the place of the version constraint, e.g. `@ncfour/projen-utils@file:../projen-utils/dist/js/projen-utils-0.0.0.jsii.tgz`. 

The changes made in #3385 to improve error messages, did not catch all valid cases of "package names" that `npm` can install. The way that change was designed, means we would have to re-implement the same algorithm npm uses. That is undesirable. 

So in this change, we partially revert the code from #3385 to instead fully rely on `npm`s install behavior, while keeping the improved error messages.

Fixes #3439

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.