feat: build workflow for sub-projects (#3317)
Creates a working `build` workflow specific to a sub-project when that project has `buildWorkflow` set to `true`.

Based on #3208. So most of the same decisions were made. The `install` command is always run in the root directory same as the `release` workflow. But based on discussion in #3304, we may want to make that configurable. I could make that change in this PR for `build`. Or just address it in a followup PR for both `build` and `release`.
I added a `buildWorkflowName` option to `NodeProject` to support customizing each build workflow name; but it still defaults to `build` or `build_{package-name}` for sub-projects.

I've successfully tested this with my own monorepo and added a few tests.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.