chore: make `bundle:task-runner` use single esbuild command (#3489)
Doesn't fix a current bug, but will help with windows support.

Instead of `bash`-isms that aren't portable (`echo ... | cat - ... | tee ... > /dev/null`) use `--banner:js=` and `--footer:js=` built-in capabilities of `esbuild`.

The resulting `lib/run-task.js` is identical in both cases.

Aside: it would be nice if we could use the work we've done in `Bundler` but alas, it doesn't give us control of the name of the task or where the output goes.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.