fix: use `bun` to print `PATH` when bun is the pm (#3490)
This is important because https://bun.sh is not only a
npm alternative, but also a node and npx alternative.

Systems with bun installed may not have node or npm, but
if `NodePackageManager.BUN` is chosen, we do know that
`bun` is available, and can print the `PATH` that way.