fix(bun): `npx` used when `bun` is package manager to print `PATH` (#3491)
This is important because https://bun.sh is not only a npm alternative, but also a node and npx alternative.

Systems with bun installed may not have node or npm, but if `NodePackageManager.BUN` is chosen, we do know that `bun` is available, and can print the `PATH` that way.

Fixes #3490 

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.