chore(docs): improve how to set up a a github app for projen  (#3516)
I found that the instructions for setting up the github actions via a github app were not quite a clear as they could have been, so raising this change to help clarify it.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.