fix(subprojects)!: conflicting projenDevDependency is added to subprojects by default (#3506)
Partially addresses: #3304

BREAKING CHANGE: by default `projen` will not be installed in subproject directories. To install `projen` in a subproject directory, explicitly set `projenDevDependency: true` in the subproject.

Installing projen in sub-projects leads to weird behavior where the sub-project is referencing different source than the parent project - which breaks things like `instanceof` expressions when the two class objects are from different sources. I think it makes more sense to only install a single copy of `projen` for an entire project graph.
I suspect that enabling sub-project installation by default was an oversight originally since it makes sense to have that default for the root project. So I think its safe to consider this a non-breaking change since the default will stay unchanged for a root project.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.