fix(release): retrying the workflow can have no effect (#3526)
When a release succeeds to create the GitHub tag, but fails on other publish steps (for example NPM), retrying the workflow has no affect. This is because we only run publishing jobs in case the tag doesn't exist. In case of a partial failure, the tag exists but the workflow didn't actually succeed and we'd like to retry.

This PR makes it so the GitHub tag is only created if all other publishing jobs succeeded. 

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.