fix(release): github release may tag the wrong commit when multiple commits are released at the same time (race condition) (#3584)
The problem, I think, is that `GITHUB_REF: ${{ github.ref }}` should be `GITHUB_REF: ${{ github.sha }}`. The ref changed during the release, so it ended up tagging the wrong commit.

Fixes #3583

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.