fix(gitlab): use WorkflowWhen in WorkflowRule (#3571)
The WorkflowRule has the type JobWhen for the when field. When in a workflow rule can be only `always` and `never`: https://docs.gitlab.com/ee/ci/yaml/#workflowrules
If used in jobs it could have more inputs: https://docs.gitlab.com/ee/ci/yaml/#when

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.