fix(jest): setting coverageProvider has no effect (#3601)
We previously forced the coverage provider as a CLI option, which prevented the value from being changed.
This PR replaces the forced CLI option with a changed default config value.
Considering this a fix since the default functionality is not changed and the option was effectively unusable before.

Also removes the check for Node 14, since this is old version is not supported by projen anymore and presumably has no projen usage.

Fixes: #3550

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.