fix: revert "fix(jest): setting coverageProvider has no effect" (#3610)
Reverts projen/projen#3601

The reverted change assumed that all users of the `Jest` component were already forced to use the `v8` coverage provider via an unchangeable CLI option. However the condition that was used to enable this behavior did not cover all cases. For anyone using custom project types not based on `NodeProject` or resetting the CLI option manually, this would have been a breaking change.

We are going to revert this PR and then re-release it as a breaking change.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.