chore: fix getting started links on landing page (#3625)
In #3606 we moved the Getting Started page to a different location.
Docusaurus as a detector for broken links, but that seems to not include HTML documents like the landing page.
So unfortunately those links did not get updated.

Fixes #3618

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.