chore(ci): require windows build to pass (#3725)
Changes the build matrix to require the Windows test to pass.
Doing this mainly by setting the `continue-on-fail` condition to false for windows build.

Also changes the semantic from "experimental" builds to a "primary" build that does extra stuff and explicitly (dis-)allowing builds to fail.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.