fix(jsii): running `projen package` on windows fails (#3723)
Fixes the failing packaging step in windows builds.

Instead of a complex bash expression, this now leverages  a feature of the projen task runner.
- Firstly, replacing the `if` statement with task step conditions 
- Secondly, use `node` to check if we are in CI. Unlike bash commands, node should always be available
- Thirdly, installing `rsync` on windows runners

Additionally this reverts the special treatment of packaging in Windows builds.

Resolves #2761

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.