chore(docs): Getting Started documentation (#3717)
# Objective
Add more details to the Getting Started documentation page.

## Overview
This PR introduces additional details to the Getting Started documentation. For consistency, please let me know if you prefer one way or the other.

I've reviewed all references to the project name `projen`. I'm following this pattern:
1. If the project title is referenced as the first word in a sentence, I am capitalizing the project name (`Projen`).
2. If used in the middle of a sentence or code comments, I'm using the lowercase version (`projen`).


As part of the updates to this page, I am introducing a number of code samples using TypeScript for demonstration purposes. The goal is to provide a more comprehensive demo of how we can personalize our projects.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.