fix(bump): exclamation operator `!` is not recognized when running bump on windows (#3745)
Fixes https://github.com/projen/projen/issues/2427

The `!` mark is used in bash because is interpreted by the `test` command. Because `test` doesn't exist in Windows cmd, then running the bump task fails. It turns out that `grep` has a flag `-v` to invert the match and works both in bash and cmd, and it solved the problem. Refactored the `publishTask.condition` to use the same expression. Added a test to check the correct handling of a release commit with message `chore(release):`.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.