feat(gitattributes)!: support endOfLine with Linux default (#3784)
Fixes https://github.com/projen/projen/issues/3545

Adds endOfLine option with LF default to avoid git mutations. This default can be overridden by later lines set by the user. It can be disabled by setting it to `{ endOfLine: EndOfLine.NONE}`, in case of a breaking change or other reason. Added tests for this new option and for other logic of `GitAttributesFile` to improve coverage. 


BREAKING CHANGE: A new gitattribute is added by default to normalize line endings (eol) of text files to LF. This prevents mutations when Windows and Linux/macOS users are working on the same project. To revert back to the previous behavior (line endings behavior is defined by user config), add the option `endOfLine: EndOfLine.NONE` to the project.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.