chore: fix sample-file.ts documentation (#3783)
Example used "source", should be "sourcePath".

This is admittedly a very minor change but it cost me a minute of debugging after using the invalid example, so I figure it's still worth it to correct this.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.