fix(release): Version class breaks with projects other than Node (#3807)
Fixes #3806

The PR https://github.com/projen/projen/pull/3801 modified the `Version` class by moving the addition of the version bump dependency from `NodeProject` to `Version`. While this makes sense intuitively, the `Version` class was previously written in a way that could be run against any project type provided the running environment had npx installed

The fix is to only add the `bumpPackage` dependency when we have a `NodePackage` available. Also adds a test and comment to prevent a future regression.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.