fix: GitHub Workflows failing due to missing hidden files (#3822)
On August 19, 2024 GitHub announced [1] that the default behavior of the `upload-artifact` action will change to not upload hidden files as they might contain sensitive information. On September 2, 2024 this change was released as a new minor version. Because projen does not lock GitHub's first-party actions to a specific release, any workflows that relied on hidden files being uploaded started failing.

As a temporary measure to unblock customers, we are reverting the action to a version. We agree with the changes GitHub has made, we just need a little more time to sort this out in projen.

Fixes #3820

[1] https://github.blog/changelog/2024-08-19-notice-of-upcoming-deprecations-and-breaking-changes-in-github-actions-runners/



---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.