fix(github): fails to upload repo patch when using `upload-artifact` >= 4.4.0 (#3819)
On August 19, 2024 GitHub announced [1] that the default behavior of the `upload-artifact` action will change to not upload hidden files as they might contain sensitive information. On September 2, 2024 this change was released as a new minor version. As a temporary measure to unblock customers, projen locked the version of the `upload-artifact` action to a previous version that is compatible with projen, see https://github.com/projen/projen/pull/3822.

This PR changes the default name of the repo patch to not use a hidden file name.
It will currently still be possible to manually specify a hidden file name for the patch file. This will be addressed when projen is switching to the latest action version.

[1] https://github.blog/changelog/2024-08-19-notice-of-upcoming-deprecations-and-breaking-changes-in-github-actions-runners/

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.