fix(jsii): build and release workflows upload full repo including hidden files (#3821)
Previously the `JsiiProject` required the complete repository to be uploaded as build artifact.
This is however not advisable due to the security risk from uploading hidden files which might contain sensitive information (see #3822).

This PR changes how the jsii build and release process works in GitHub Actions. Instead of uploading the full repository in its post-build state, we only store the js package as artifact. The restore the post-build state, we can do a checkout of the same commit and extract all files from the js package. This technically doesn't cover _all_ of the repo in post-build state. It still works because only packaged files are relevant for release artifacts.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.