fix(typescript): can't release when package is disabled (#3847)
Fixes #3846

The `Release` component of `NodeProject` requires the packaging task to be available and functioning. Previously it was possible to configure the project so that `release: true` and `package: false`, causing a failure during the release. This was most prominent in the `TypeScriptAppProject`, which is an opinionated TypeScript project type that basically indicates that the package should not be published. Because it has both `release` and `package` disabled by default, it's easy to just enable releases but not packaging.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.