feat(node)!: default to LTS version in GitHub Actions workflows and match `@types/node` to TypeScript version (#3855)
Previously the default node version used in GitHub Actions workflows was explicitly defined by projen and updated periodically. A similar change has been made to the version of `@types/node` for typescript projects. With these changes projen won't have to release an update anymore to update the versions. Instead a sensible default is used that will automatically move forward. If users desire more explicit control over versions, they can opt-in to control these versions themselves.

Also clarifies the intended usage of `minNodeVersion` and `maxNodeVersion` which should not be used in most cases.

BREAKING CHANGE: The `workflowNodeVersion` used in GitHub Actions Workflows now defaults to `lts/*`. To revert to the previous behavior, set `workflowNodeVersion: '18.x'`. 

BREAKING CHANGE: The version of the`@types/node` dev dependency now falls back to match the version of TypeScript and then the latest version if `minNodeVersion` is not set. To revert to the previous behavior, add an explicit devDependency on `@types/node` with the desired version.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.