fix: bundledDependencies are silently ignored by NPM
Certain versions of npm (at least my tested version, `10.8.2`) silently
refuse to bundle a dependency if it doesn't occur in `dependencies`, and
`dependencies` only. Specifically, if it also occurs in
`devDependencies`, `npm pack` will just silently not bundle the
directory.

Add a check for that to prevent users from falling into this trap.

Also add a clarification on the `peerDependencies` check. It makes
sense, but it took me longer than should be necessary to understand,
and I like the computer to explain things to my peanut-sized brain 😉.