feat(gitlab): add path option to CiConfiguration (#3882)
This introduces a `path` option to `GitlabConfiguration` and `NestedConfiguration`, giving users the option to produce Gitlab CI configuration and Gitlab CI templates  in the files of their choosing. This is useful for a pattern where we only want to execute a job if known input files or the job definition itself has changed, consider:

```
├── gitlab-ci.yml
├── a
│   ├── job.yml
│   └── input-a
└── b
    ├── job.yml
    └── input-b
```

```
# b.yml
a:
  script: []
  rules:
     changes: b/**/*
b:
  script: []
  rules:
     changes: b/**/*
```

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.