feat: default the AWS CDK CLI version to "latest"
projen currently installs the same CDK CLI version as the requested CDK
framework version. The CDK requirement used to be that `CLI Version >=
Lib Version`, and setting them to the same value was a simple way of
satisfying that requirement.

In the near future, the version of the CDK CLI and the CDK Library
will not be the same anymore (see
https://github.com/aws/aws-cdk/issues/32775). At that point, this
logic will start to fail.

Instead, change the default to `aws-cdk@latest`. This still satisfies
the requirement that `CLI Version >= Lib Version` (for old CDK
releases), and more importantly also satisfies the requirement that `CDK
Release Date >= Lib Release Date` (for new CDK releases).

A new optional parameter is added to specifically configure the CLI version
if desired.