feat: ReleaseTrigger.workflowDispatch
I want to be able to trigger a release only by starting a workflow on
GitHub, and nothing else.

The `manual` trigger implies a task I need to run from a working copy,
but I want just the scheduled job, except it runs whenever I choose
and not on an actual schedule.