feat(release): `ReleaseTrigger.workflowDispatch` (#4081)
* feat: ReleaseTrigger.workflowDispatch

I want to be able to trigger a release only by starting a workflow on
GitHub, and nothing else.

The `manual` trigger implies a task I need to run from a working copy,
but I want just the scheduled job, except it runs whenever I choose
and not on an actual schedule.

* chore: self mutation

Signed-off-by: github-actions <github-actions@github.com>

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>