fix(bump): output of nextVersionCommand is ignored if there are 0 commits
The `nextVersionCommand` script should have full control over whether a
version bump gets executed or not, but currently there is logic that
decides that if there are 0 commits since the last tag, the bump will
always be skipped.

Change that to take the output of the `nextVersionCommand` script into
account.