chore: add helper to convert arbitrary JSON to GitHub expressions
> This is extracted from another PR where I ended up changing the
structure so I don't need it anymore, but I think a function like
this is useful in general so I want to commit it anyway.

Converting arbitrary data to GitHub Actions format reliably is not super
hard but it's not extremely trivial either, especially once we get into
things like quoting strings and transferring data like arrays.

This adds a helper function that takes away the guesswork and just
does the conversion.