feat: restrict branches on which we enable autoqueue
The ability to only enable autoqueue'ing for PRs targeting the `main`
branch–and crucially: also enable autoqueuing when the PR is retargeted
from whatever its current base is to the `main` branch–allows automatic
merging of a set of stacked PRs in the right order.

Example:

- Create PR #1 from branch `a`, targeting `main`.
- Create PR #2 from branch `b`, targeting branch `a`.
- Create PR #3 from branch `c`, targeting branch `b`.

Initially, PR #1 will be set to auto-merge, PRs #2 and #3 will not.

Once PR #1 passes all of its requirements it will merge. That will delete
branch `a` and change  the target branch of PR #2 change to `main`. At that
point, auto-queueing will switch on for PR #2 and it gets merged, etc.