feat: `nextVersionCommand` can filter bump decision
Previously, when `nextVersionCommand` was specified it would need to
entirely determine the bump decision, or it was not specified and the
bump decision was left to `commit-and-tag-version`.

This makes `nextVersionCommand` need to do a bunch of work that it
doesn't necessarily want to do, and it's not in a great place to make
those kinds of decisions.

In this PR, change the process to be the following:

- `commit-and-tag-version` determines the bump type (using its
  `--dry-run` feature)
- We then pass the decision into `nextVersionCommand` via a new
  environment variable named `$SUGGESTED_BUMP`.
- The script is free to return a new bump type, or return nothing. If
  nothing is returned, the `$SUGGESTED_BUMP` type is still used.
- Finally the selected bump type is applied.

This gives the script the ability to reuse decisions made by
`commit-and-tag-version`, while remaining backwards compatible.