fix: releasing sometimes tags wrong commit
It looks like `gh release create --target $SHA` sometimes creates releases
that point to the wrong commit. This seems to happen if the `main`
branch has updated since the release process has started: in that case,
the newest commit on `main` gets tagged, instead of the one that
started the workflow.

(This happens for release workflows that are started using a
`workflow_dispatch` command -- releases that run continously have
additional protections built in to abort themselves if the `main`
branch has shifted).

We're not quite sure what is causing this, but one way of reducing
the uncertainty here is doing the tagging ourselves manually (instead
of having GitHub create the tag as part of creating the release)
and passing the `--verify-tag` flag to the `gh release` command
which requires that the tag already exists.

Fixes https://github.com/projen/projen/issues/4176