fix: releasing tags wrong commit when run via workflow_dispatch (#4188)
When running our release command, we run

```
gh release create --target $GITHUB_REF ...
```

And set:

```
env:
  GITHUB_REF: {{ github.sha }}
```

This *looks* like it should tag the release to a given commit.

However, `GITHUB_REF` (and in fact, `GITHUB_REPOSITORY` and
`GITHUB_TOKEN`) are special and automatically set by GitHub,
overwriting whatever we have configured.

For a `push` event `GITHUB_REF` contains the SHA that started the
workflow, but for a `workflow_dispatch` event `GITHUB_REF` contains
`refs/heads/main`, because that's the branch we got started on.

That means that if the `main` branch changes while the `release`
workflow is running, ultimately the wrong commit gets tagged.

Instead of trying to assign `{{ github.sha }}` to `GITHUB_REF`,
we can use the predefined environment variable `GITHUB_SHA`
to achieve the effect we need.

Fixes https://github.com/projen/projen/issues/4176