fix(github): built-in outputs are not working cross-platform (#4273)
All default workflows that set GitHub Actions outputs do not work
cross-platform. This is because outputs are set by assigning a value to
a "magic" environment variable, and that assignment needs to be
different depending on the shell that is used. See
https://github.com/orgs/community/discussions/39715 for more details.

To ensure the workflows also work on Windows runners by default, we are
adding `shell: bash` to all workflow steps that write outputs.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>