chore: move CI workloads to D: drive on windows build to improve performance (#4288)
Build times on Windows have crept up to about an hour and roughly 5x
slower compared to the Ubuntu build. There is literature [1] that
windows build on GitHub have bad I/O performance and that the D: drive
should be faster.

This change moves the temp directory and the yarn cache to the D: drive.
Overall this reduces the build time by 40 min down to 17 min (compared
to 11 min on Ubuntu). So much better!


[1]
https://chadgolden.com/blog/github-actions-hosted-windows-runners-slower-than-expected-ci-and-you

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.