feat(release): support release environment in GitHub Actions (#4309)
This feature allows adding an environment to the GitHub Actions job
that's handling the release. This environment can then be used to add an
explicit approval step to the release or limit who can initiate a
release through environment protection rules.

The environment can be set at a package level, or to a different
environment for a release branch, or for projects types that publish
multiple artifacts per artifact.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.