fix: special characters in branch name interfere with shell parsing (#4365)
In the self-mutation workflow, we were passing the Git ref name through
an environment variable as recommended by GitHub, but we still need to
quote it to make sure any special characters in there are defused by the
shell.

Turn:

```
          git push origin HEAD:$PULL_REQUEST_REF
```

into

```
          git push origin "HEAD:$PULL_REQUEST_REF"
```

Since this is a `pull_request` workflow, there is no ability for a
remote attacker to compromise a repository. Only someone with write
access to the repository can craft a branch name that will inject shell
commands; at which point they can do worse already.

Nevertheless, we are fixing this escaping issue since it will be flagged
by certain scanner tools.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.