feat(awscdk)!: AutoDiscover Lambda Runtime defaulting to nodejs22.x (#4382)
BREAKING CHANGE: The default Lambda Runtime is updated to `nodejs22.x`,
as the current default runtime is `nodejs18.x` which is already
deprecated (see:
[here](https://docs.aws.amazon.com/lambda/latest/dg/lambda-runtimes.html#runtimes-deprecated)).
To keep the previous version, set
[`lambdaOptions.runtime`](https://projen.io/docs/api/awscdk?_highlight=runtime#projen.awscdk.LambdaFunctionCommonOptions.property.runtime)
to the desired runtime.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

Co-authored-by: Rafał Pawłaszek <rpawlasz@amazon.com>