fix: `Jest` component cannot be created as a child of non-`NodeProject` (#4101)
The fix is just a proxy to introduce a new helper function
`closestProjectMustBe`.

The helper allows components to easily ensure they are created within a
project of a certain type and correctly declare the types of its
enclosing project. This approach is more flexible than requiring a
specific project type inside the class constructor, since the component
can now be created nested inside an abstraction. If a project of the
required type is not found, an error is thrown.

An implementing component would use this like so:

```ts
export class MyComponent extends Component {
  public readonly project: NodeProject;
  constructor(scope: IConstruct, options: JestOptions = {}) {
    super(scope);
    this.project = closestProjectMustBe(scope, NodeProject, new.target.name);
  }
}
```

---

I have also refactored the existing `findClosestProject()` function to
explicitly take the name of the `constructInCreation` instead of
implicitly calling `new.target.name`. This was to align it with the new
helper. Explicitness is better here, since otherwise the function can
only be used inside class constructors.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.